Move CustomSyncServerSettingsFragment to a new file
